simulators/requirements.txt
databricks-sdk>=0.45.0
databricks-zerobus-sdk>=0.2.0
numpy>=1.26
python-dateutil>=2.8
click>=8.1
